% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6-classes.R
\docType{class}
\name{Adapter}
\alias{Adapter}
\title{R6 class that selects a HPDS-hosted resources on a PIC-SURE network.}
\format{\code{PicSureHpdsAdapter} object.}
\usage{
Adapter
}
\value{
Object of \code{\link{R6Class}} with various utility methods for PIC-SURE network and to access a HPDS-hosted resource.
}
\description{
R6 class that selects a HPDS-hosted resources on a PIC-SURE network.
}
\section{Methods}{

\describe{
  \item{Documentation}{For full documentation of each method go to https://github.com/hms-dbmi/pic-sure-r-adapter-hpds}
  \item{\code{new(PicSureConnection)}}{This method is used to create new object of this class which uses the passed PicSureConnection object for communication with the PIC-SURE network.}

  \item{\code{version()}}{This method prints the current package version.}
  \item{\code{list()}}{This method prints a list of UUIDs of all resources hosted by the currently connected PIC-SURE network.}
  \item{\code{useResource(resource_uuid)}}{This method returns a new \code{PicSureHpdsResourceConnection} object configured to connect the specified HPDS-hosted resource on the PIC-SURE Network identified by the given \code{resource_uuid}.}}
}

\keyword{data}
