% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{addClause}
\alias{addClause}
\title{Add a clause to a query}
\usage{
addClause(
  query,
  keys,
  type = "FILTER",
  min = NULL,
  max = NULL,
  categories = NULL,
  lookupVariablesOverride = NULL
)
}
\arguments{
\item{query}{A query object}

\item{keys}{A key or list of keys to add a clause for}

\item{type}{One of: (“SELECT”, “REQUIRE”, “ANYOF”, “FILTER”)}

\item{min}{Optional parameter for continuous "FILTER" clauses. Specifies the minimum value}

\item{max}{Optional parameter for continuous "FILTER" clauses. Specifies the maximum value}

\item{categories}{Optional parameter for categorical "FILTER" clauses to match}
}
\value{
A query object
}
\description{
Add a clause to a query
}
\examples{

# query <- picsure::newQuery(session)
# query <- picsure::addClause(query, "\\demographics\\AGE\\", max = 10)

}
