% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc.R
\name{bdc.getQueryByUUID}
\alias{bdc.getQueryByUUID}
\title{Re-creates a query object from a queryUUID}
\usage{
bdc.getQueryByUUID(session, queryUUID)
}
\arguments{
\item{session}{Current PIC-SURE session}

\item{queryUUID}{The query UUID as copied from PIC-SURE}
}
\description{
Re-creates a query object from a queryUUID
}
\examples{

# query <- picsure::bdc.getQueryByUUID(session, query = "your query uuid")

}
