% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc.R
\name{bdc.newQuery}
\alias{bdc.newQuery}
\title{Creates a query object to query a PIC-SURE instance}
\usage{
bdc.newQuery(session)
}
\arguments{
\item{session}{Current PIC-SURE session}
}
\value{
A query object
}
\description{
Creates a query object to query a PIC-SURE instance
}
\examples{

# query <- picsure::bdc.newQuery(session)

}
