% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc.R
\name{bdc.searchPicsure}
\alias{bdc.searchPicsure}
\title{Performs a search of variables in PIC-SURE for a given keyword}
\usage{
bdc.searchPicsure(
  session,
  keyword = "",
  resultType = "DATA_FRAME",
  includeValues = FALSE,
  limit = 0,
  offset = 0
)
}
\arguments{
\item{session}{Current session}

\item{keyword}{Keyword to search for}

\item{resultType}{Optional parameter to specify result type: Valid values:
DATA_FRAME (default) which converts the response to a data frame
DICTIONARY which returns a list
VARIABLE_PATHS which returns the unique variable paths}

\item{includeValues}{Boolean to enable including all values with categorical
variables. Default is FALSE. Use with caution, this can be a large amount of data}

\item{limit}{Optonal, default 0, use to limit results. Returns the first N results}

\item{offset}{Optional, default 0, use to paginate results. Returns results starting at N}
}
\value{
A PIC-SURE session
}
\description{
Performs a search of variables in PIC-SURE for a given keyword
}
\examples{

searchResults <- picsure::bdc.searchPicsure(session, "heart", resultType = "DATA_FRAME")

}
