% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6-classes.R
\docType{class}
\name{BypassAdapter}
\alias{BypassAdapter}
\title{R6 class that selects a HPDS-hosted resources directly without the use of the PIC-SURE Connection library.}
\format{
\code{PicSureHpdsBypassAdapter} object.
}
\value{
Object of \code{\link{R6Class}} used to access a HPDS-hosted resource without using a PIC-SURE network.
}
\description{
R6 class that selects a HPDS-hosted resources directly without the use of the PIC-SURE Connection library.

R6 class that selects a HPDS-hosted resources directly without the use of the PIC-SURE Connection library.
}
\section{Methods}{

\describe{
  \item{Documentation}{For full documentation of each method go to https://github.com/hms-dbmi/pic-sure-r-adapter-hpds}
  \item{\code{new(url, token)}}{This method is used to create new object of this class which connects directly to a HPDS-hosted resource.}

  \item{\code{list()}}{This method prints a list of UUIDs of all resources hosted by the currently connected PIC-SURE network.}
  \item{\code{useResource(resource_uuid)}}{This method returns a new \code{PicSureHpdsResourceConnection} object configured to connect the specified HPDS-hosted resource on the PIC-SURE Network identified by the given \code{resource_uuid}.}}
}

\keyword{data}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{BypassAdapter$new()}}
\item \href{#method-useResource}{\code{BypassAdapter$useResource()}}
\item \href{#method-clone}{\code{BypassAdapter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BypassAdapter$new(url_arg, token_arg = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-useResource"></a>}}
\if{latex}{\out{\hypertarget{method-useResource}{}}}
\subsection{Method \code{useResource()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BypassAdapter$useResource(resource_uuid)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BypassAdapter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
