% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{deleteClause}
\alias{deleteClause}
\title{Deletes all clauses from this query for a given key}
\usage{
deleteClause(query, key)
}
\value{
The updated query
}
\description{
Deletes all clauses from this query for a given key
}
