% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{exportTSV}
\alias{exportTSV}
\title{Export query results to a TSV file.}
\usage{
exportTSV(session, query, path, ...)
}
\arguments{
\item{session}{A session object produced by [`connect()`][picsure::connect].}

\item{query}{A clause group from
[`buildClauseGroup()`][picsure::buildClauseGroup].}

\item{path}{Destination file path.}

\item{...}{Additional keyword arguments forwarded to the Python
`Session.exportPFB()` call.}
}
\value{
The path, invisibly.
}
\description{
Export query results to a TSV file.
}
\examples{
\dontrun{
picsure::exportTSV(bdc, full_query, "~/cohort.tsv")
}
}
