% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpds.R
\name{extract.entries}
\alias{extract.entries}
\title{Extract the unique keys of all the results in a given data dictionary lookup.}
\usage{
extract.entries(dictionary.results, verbose = FALSE)
}
\arguments{
\item{dictionary.results}{A dictionary results object.}

\item{verbose}{Flag to display additional runtime information.}
}
\value{
A collection of result entries.
}
\description{
Extract the unique keys of all the results in a given data dictionary lookup.
}
