% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpds.R
\name{extract.keys}
\alias{extract.keys}
\title{Extract the unique keys of all the results in a given data dictionary lookup.}
\usage{
extract.keys(dictionary.results, verbose = FALSE)
}
\arguments{
\item{dictionary.results}{A data dictionary search results object.}

\item{verbose}{Flag to display additional runtime information.}
}
\value{
A list of unique keys for all the data dictionary search results.
}
\description{
Extract the unique keys of all the results in a given data dictionary lookup.
}
