% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{generateQueryJSON}
\alias{generateQueryJSON}
\title{Generate JSON representation of the query to be sent to the PIC-SURE api. This handles various quirks
between the required JSON schema and how jsonlite serializes R data types}
\usage{
generateQueryJSON(query, expectedResultType)
}
\description{
Generate JSON representation of the query to be sent to the PIC-SURE api. This handles various quirks
between the required JSON schema and how jsonlite serializes R data types
}
