% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\docType{data}
\name{GenomicFilterKey}
\alias{GenomicFilterKey}
\title{Genomic annotation keys for `buildGenomicFilter()`.}
\format{
A list of `picsure_enum_member` objects:
\describe{
  \item{`GENE_WITH_VARIANT`}{Gene affected by a variant.}
  \item{`VARIANT_CONSEQUENCE_CALCULATED`}{Calculated variant consequence.}
  \item{`VARIANT_FREQUENCY_AS_TEXT`}{Population-frequency bucket (see
    [`VariantFrequency`][picsure::VariantFrequency]).}
  \item{`VARIANT_CLASS`}{Variant class.}
  \item{`VARIANT_SEVERITY`}{Virtual severity key (see
    [`VariantSeverity`][picsure::VariantSeverity]).}
}
}
\usage{
GenomicFilterKey
}
\description{
Pass a member to [`buildGenomicFilter()`][picsure::buildGenomicFilter]'s
`key` argument, or the equivalent string (validated by the Python adapter).
Mirrors Python's `picsure.GenomicFilterKey`.
}
\details{
`VARIANT_SEVERITY` is a virtual key: the builder expands a
[`VariantSeverity`][picsure::VariantSeverity] bucket into the matching
`Variant_consequence_calculated` values.
}
\examples{
\dontrun{
picsure::buildGenomicFilter(
  picsure::GenomicFilterKey$GENE_WITH_VARIANT,
  values = c("BRCA1")
)
}
}
\keyword{datasets}
