% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpds.R
\name{get.full.dictionary}
\alias{get.full.dictionary}
\title{Import the entire data dictionary of a PIC-SURE resource.}
\usage{
get.full.dictionary(resource, useQueryScope = TRUE)
}
\arguments{
\item{resource}{A PIC-SURE resource object.}

\item{useQueryScope}{A boolean indicating if the resource's data dictionary should be filtered by the user's queryScopes variable.}
}
\value{
A data frame containing the data dictionary records. Phenotype variables may have been filtered by queryScopes.
}
\description{
Import the entire data dictionary of a PIC-SURE resource.
}
\examples{

# myconn <- picsure::connect(url="http://your.server/PIC-SURE/", token="your-security-token")
# myres <- hpds::get.resource(connection=myconn, resourceUUID="YOUR-UUID-0000")
# all.terms.df <- hpds::get.full.dictionary(resource=myres, useQueryScope=FALSE)

}
