% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpds.R
\name{get.resource}
\alias{get.resource}
\title{Get a new instance reference of a PIC-SURE resource.}
\usage{
get.resource(connection, resourceUUID, verbose = FALSE)
}
\arguments{
\item{connection}{A PIC-SURE connection object.}

\item{resourceUUID}{The UUID identity of a Resource hosted via the PIC-SURE connection.}

\item{verbose}{Flag to display additional runtime information.}
}
\value{
An object which provides access to the requested Resource.
}
\description{
Get a new instance reference of a PIC-SURE resource.
}
