% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{getDataFrame}
\alias{getDataFrame}
\title{Maps the results of a search query to a phenotypes data frame, and an info data frame}
\usage{
getDataFrame(results)
}
\description{
Maps the results of a search query to a phenotypes data frame, and an info data frame
}
