% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/picsure.R
\name{getGenomicAnnotations}
\alias{getGenomicAnnotations}
\title{Returns the genomic annotations for this session, as a data frame}
\usage{
getGenomicAnnotations(session)
}
\arguments{
\item{session}{Current PIC-SURE session}
}
\description{
Returns the genomic annotations for this session, as a data frame
}
