% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/picsure.R
\name{getGenotypeAnnotations}
\alias{getGenotypeAnnotations}
\title{Returns the genotype annotations for this session, as a data frame}
\usage{
getGenotypeAnnotations(session)
}
\arguments{
\item{session}{Current PIC-SURE session}
}
\description{
Returns the genotype annotations for this session, as a data frame
}
