% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/picsure.R
\name{getJSON}
\alias{getJSON}
\title{Make a GET request to the PIC-SURE service. Handles setting content type, and required headers. Throws an exception if there are
any errors, or if the response code is not 200.}
\usage{
getJSON(session, path, psama = FALSE)
}
\arguments{
\item{session}{The current PIC-SURE session}

\item{path}{A url path to make a GET request to}
}
\value{
the result of the request, deserialized from JSON using jsonlite
}
\description{
Make a GET request to the PIC-SURE service. Handles setting content type, and required headers. Throws an exception if there are
any errors, or if the response code is not 200.
}
