% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{getQueryByUUID}
\alias{getQueryByUUID}
\title{Builds a query with the given query UUID}
\usage{
getQueryByUUID(session, queryUUID)
}
\arguments{
\item{session}{The current session object}

\item{queryUUID}{The query UUID as copied from PIC-SURE}
}
\value{
A initialized query object that can be modified
}
\description{
Builds a query with the given query UUID
}
\examples{

# query <- picsure::getQueryByUUID(session, "a2ba4b22-f85b-4388-a098-6c72cd55e9d3")

}
