% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/picsure.R
\name{getResources}
\alias{getResources}
\title{Returns a list of resources for this session}
\usage{
getResources(session)
}
\arguments{
\item{session}{Current PIC-SURE session}
}
\description{
Returns a list of resources for this session
}
