% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{getResultByQueryUUID}
\alias{getResultByQueryUUID}
\title{Executes a query with a given queryUUID}
\usage{
getResultByQueryUUID(session, queryUUID)
}
\arguments{
\item{session}{The current session object}

\item{queryUUID}{The unique id of the query to run}
}
\value{
The result of the query as a data frame
}
\description{
Executes a query with a given queryUUID
}
\examples{

# result <- picsure::getResultByQueryUUID(session, "e10882ae-542a-4e38-ae28-399e13ac38c1")

}
