% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpds.R
\docType{package}
\name{hpds}
\alias{hpds}
\alias{hpds-package}
\title{hpds: An Adapter Library for HPDS Services on PIC-SURE Networks.}
\description{
This package is used to interact with HPDS Services and data hosted
on a PIC-SURE network.  This library requires the 'picsure' package
to handle all connectivity and authentication with the PIC-SURE
Network itself.
}
