% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/picsure.R
\name{initializeSession}
\alias{initializeSession}
\title{Creates a session to access a PIC-SURE instance}
\usage{
initializeSession(url, token, psama_url = FALSE, getDictionary = NULL)
}
\arguments{
\item{url}{Url to a PIC-SURE instance}

\item{token}{A user token to access this instance. This can be found in the "User Profile" tab
on the PIC-SURE site}

\item{psama_url}{(Optional) URL to override the default PSAMA endpoint}
}
\value{
An object which provides access to a PIC-SURE instance
}
\description{
Creates a session to access a PIC-SURE instance
}
\examples{

# session <- picsure::initializeSession(url="http://your.server/PIC-SURE/", token="your-security-token")

}
