% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{lookupGenomicVariables}
\alias{lookupGenomicVariables}
\title{Gets a genomic variable from the session dictionary, if it exists}
\usage{
lookupGenomicVariables(query, keys)
}
\description{
Gets a genomic variable from the session dictionary, if it exists
}
