% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{lookupVariables}
\alias{lookupVariables}
\title{Gets a variable from the session dictionary, if it exists}
\usage{
lookupVariables(query, keys)
}
\description{
Gets a variable from the session dictionary, if it exists
}
