% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{mapInfoResult}
\alias{mapInfoResult}
\title{Maps a row from the info result to a row to be included in the genomic annotations dataframe}
\usage{
mapInfoResult(result)
}
\description{
Maps a row from the info result to a row to be included in the genomic annotations dataframe
}
