% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{mapPhenotypeResult}
\alias{mapPhenotypeResult}
\title{Maps a row from the phenotype result to a row to be included in the dictionary dataframe}
\usage{
mapPhenotypeResult(result)
}
\description{
Maps a row from the phenotype result to a row to be included in the dictionary dataframe
}
