% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpds.R
\name{new.query}
\alias{new.query}
\title{Create a new query instance with no restrictions.}
\usage{
new.query(resource, verbose = FALSE)
}
\arguments{
\item{resource}{A resource object that the returned query object will execute against.}

\item{verbose}{Flag to display additional runtime information.}
}
\description{
Create a new query instance with no restrictions.
}
\examples{

# myconn <- picsure::connect(url="http://your.pic-sure.server/PIC-SURE/", token="your-security-token")
# myres <- hpds::get.resource(connection=myconn, resourceUUID="612d00b7-YOUR-UUID-0000-1efe587f0bcd")
# myquery <- hpds::new.query(resource=myres)

}
