% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{newQuery}
\alias{newQuery}
\title{Creates a query object to query a PIC-SURE instance}
\usage{
newQuery(session)
}
\arguments{
\item{session}{Current PIC-SURE session}
}
\value{
A query object
}
\description{
Creates a query object to query a PIC-SURE instance
}
\examples{

# query <- picsure::newQuery(session)

}
