% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{parseQueryTemplate}
\alias{parseQueryTemplate}
\title{Parses the query template and updates appropriate query fields. This may be incomplete,
but currently handles known query templates}
\usage{
parseQueryTemplate(query)
}
\description{
Parses the query template and updates appropriate query fields. This may be incomplete,
but currently handles known query templates
}
