% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6-classes.R
\docType{class}
\name{PicSureHpdsBypassConnection}
\alias{PicSureHpdsBypassConnection}
\title{R6 class that is a code-shim used by the PicSureBypassAdapter - DO NOT USE THIS OBJECT DIRECTLY!}
\format{
\code{\link{PicSureHpdsBypassConnection}} object.
}
\value{
Object of \code{\link{R6Class}} used by the \code{PicSureBypassAdapter} to access a HPDS-hosted resource without using a PIC-SURE network.
}
\description{
R6 class that is a code-shim used by the PicSureBypassAdapter - DO NOT USE THIS OBJECT DIRECTLY!

R6 class that is a code-shim used by the PicSureBypassAdapter - DO NOT USE THIS OBJECT DIRECTLY!
}
\section{Methods}{

\describe{
  \item{Documentation}{For full documentation of each method go to https://github.com/hms-dbmi/pic-sure-r-adapter-hpds}}
}

\keyword{data}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PicSureHpdsBypassConnection$new()}}
\item \href{#method-list}{\code{PicSureHpdsBypassConnection$list()}}
\item \href{#method-INTERNAL_api_obj}{\code{PicSureHpdsBypassConnection$INTERNAL_api_obj()}}
\item \href{#method-getResources}{\code{PicSureHpdsBypassConnection$getResources()}}
\item \href{#method-clone}{\code{PicSureHpdsBypassConnection$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PicSureHpdsBypassConnection$new(url_arg, token_arg)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list"></a>}}
\if{latex}{\out{\hypertarget{method-list}{}}}
\subsection{Method \code{list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PicSureHpdsBypassConnection$list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-INTERNAL_api_obj"></a>}}
\if{latex}{\out{\hypertarget{method-INTERNAL_api_obj}{}}}
\subsection{Method \code{INTERNAL_api_obj()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PicSureHpdsBypassConnection$INTERNAL_api_obj()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getResources"></a>}}
\if{latex}{\out{\hypertarget{method-getResources}{}}}
\subsection{Method \code{getResources()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PicSureHpdsBypassConnection$getResources()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PicSureHpdsBypassConnection$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
