% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6-classes.R
\docType{class}
\name{PicSureHpdsQuery}
\alias{PicSureHpdsQuery}
\title{R6 class used to build a multi-use query to search against a HPDS resource's data - DO NOT CREATE THIS OBJECT DIRECTLY!}
\format{\code{\link{PicSureHpdsQuery}} object.}
\usage{
PicSureHpdsQuery
}
\value{
Object of \code{\link{R6Class}} used to access a HPDS-hosted resource's data dictionary.
}
\description{
R6 class used to build a multi-use query to search against a HPDS resource's data - DO NOT CREATE THIS OBJECT DIRECTLY!
}
\section{Methods}{

\describe{
  \item{Documentation}{For full documentation of each method go to https://github.com/hms-dbmi/pic-sure-r-adapter-hpds}
  \item{\code{new(connection)}}{This method is used to create new object of this class. DO NOT CREATE THIS OBJECT DIRECTLY!}

  \item{\code{show()}}{This method displays a list of all settings specified for the query.}
  \item{\code{select()}}{This method accesses a reference to a query parameter list.}
  \item{\code{crosscounts()}}{This method accesses a reference to a query parameter list.}
  \item{\code{require()}}{This method accesses a reference to a query parameter list.}
  \item{\code{anyof()}}{This method accesses a reference to a query parameter list.}
  \item{\code{filter()}}{This method accesses a reference to a query parameter list.}
  \item{\code{getCount()}}{This method returns a count of how many records are discovered by the query.}
  \item{\code{getResults()}}{This method returns the records discovered by the query.}
  \item{\code{getResultsDataFrame()}}{This method returns the discovered records in a dataframe format.}
  \item{\code{getRunDetails()}}{This method returns information the performance of the query.}}
}

\keyword{data}
