% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6-classes.R
\docType{class}
\name{PicSureHpdsResourceConnection}
\alias{PicSureHpdsResourceConnection}
\title{R6 class that allows access to the data dictionary and query services of a selected HPDS-hosted resources on a PIC-SURE network.}
\format{
\code{\link{PicSureHpdsResourceConnection}} object.
}
\value{
Object of \code{\link{R6Class}} used to access the dictionary and query services via the objects it returns.
}
\description{
R6 class that allows access to the data dictionary and query services of a selected HPDS-hosted resources on a PIC-SURE network.

R6 class that allows access to the data dictionary and query services of a selected HPDS-hosted resources on a PIC-SURE network.
}
\section{Methods}{

\describe{
  \item{Documentation}{For full documentation of each method go to https://github.com/hms-dbmi/pic-sure-r-adapter-hpds}
  \item{\code{new(connection, resource_uuid)}}{This method is used to create new object of this class which uses the passed PicSureConnection object for communication with the PIC-SURE network along with a UUID to identify a HPDS-hosted resource.}

  \item{\code{dictionary()}}{This method returns a \code{PicSureHpdsDictionary} object which is used to run lookups against a resources data dictionary.}
  \item{\code{query()}}{This method returns a new \code{PicSureHpdsQuery} object configured to run all commands against the previously specified HPDS-hosted resource.}}
}

\keyword{data}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PicSureHpdsResourceConnection$new()}}
\item \href{#method-version}{\code{PicSureHpdsResourceConnection$version()}}
\item \href{#method-dictionary}{\code{PicSureHpdsResourceConnection$dictionary()}}
\item \href{#method-query}{\code{PicSureHpdsResourceConnection$query()}}
\item \href{#method-retrieveQueryResults}{\code{PicSureHpdsResourceConnection$retrieveQueryResults()}}
\item \href{#method-getQueryByUUID}{\code{PicSureHpdsResourceConnection$getQueryByUUID()}}
\item \href{#method-clone}{\code{PicSureHpdsResourceConnection$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PicSureHpdsResourceConnection$new(connection, resource_uuid)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-version"></a>}}
\if{latex}{\out{\hypertarget{method-version}{}}}
\subsection{Method \code{version()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PicSureHpdsResourceConnection$version()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dictionary"></a>}}
\if{latex}{\out{\hypertarget{method-dictionary}{}}}
\subsection{Method \code{dictionary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PicSureHpdsResourceConnection$dictionary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-query"></a>}}
\if{latex}{\out{\hypertarget{method-query}{}}}
\subsection{Method \code{query()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PicSureHpdsResourceConnection$query(loadQuery = NA)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-retrieveQueryResults"></a>}}
\if{latex}{\out{\hypertarget{method-retrieveQueryResults}{}}}
\subsection{Method \code{retrieveQueryResults()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PicSureHpdsResourceConnection$retrieveQueryResults(query_uuid = NA)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getQueryByUUID"></a>}}
\if{latex}{\out{\hypertarget{method-getQueryByUUID}{}}}
\subsection{Method \code{getQueryByUUID()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PicSureHpdsResourceConnection$getQueryByUUID(query_uuid)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PicSureHpdsResourceConnection$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
