% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/picsure.R
\name{postJSON}
\alias{postJSON}
\title{Make a POST request to the PIC-SURE service. Handles setting content type, and required headers. Throws an exception if there are
any errors, or if the response code is not 200.}
\usage{
postJSON(session, path, body, responseDeserializer = deserializeJSON)
}
\arguments{
\item{session}{The current PIC-SURE session}

\item{path}{A url path to POST to}

\item{body}{A string of JSON data to post to the service}

\item{responseDeserializer}{A function to deserialize the data returned. Default is deserializeJSON}
}
\value{
a string containing the response from the request, or the result of responseDeserializer, if provided
}
\description{
Make a POST request to the PIC-SURE service. Handles setting content type, and required headers. Throws an exception if there are
any errors, or if the response code is not 200.
}
