% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpds.R
\name{query.anyof.add}
\alias{query.anyof.add}
\title{Add an "AnyOf" restriction from a query instance.}
\usage{
query.anyof.add(query, keys, verbose = FALSE)
}
\arguments{
\item{query}{A query instance object.}

\item{keys}{One or more keys to add to the given query object's AnyOf list.}

\item{verbose}{Flag to display additional runtime information.}
}
\description{
Add an "AnyOf" restriction from a query instance.
}
