% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpds.R
\name{query.anyof.delete}
\alias{query.anyof.delete}
\title{Delete an "AnyOf" restriction from a query instance.}
\usage{
query.anyof.delete(query, keys, verbose = FALSE)
}
\arguments{
\item{query}{A query instance object.}

\item{keys}{One or more keys to delete from the given query object's AnyOf list.}

\item{verbose}{Flag to display additional runtime information.}
}
\description{
Delete an "AnyOf" restriction from a query instance.
}
