% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpds.R
\name{query.crosscounts.delete}
\alias{query.crosscounts.delete}
\title{Delete a "cross-counts" restriction from a query instance.}
\usage{
query.crosscounts.delete(query, keys, verbose = FALSE)
}
\arguments{
\item{query}{A query instance object.}

\item{keys}{One or more keys to delete from the given query object's cross-count list.}

\item{verbose}{Flag to display additional runtime information.}
}
\description{
Delete a "cross-counts" restriction from a query instance.
}
\examples{

# myconn <- picsure::connect(url="http://your.pic-sure.server/PIC-SURE/", token="your-security-token")
# myres <- hpds::get.resource(connection=myconn, resourceUUID="612d00b7-YOUR-UUID-0000-1efe587f0bcd")
# myquery <- hpds::new.query(resource=myres)
# hpds::query.crosscounts.add(query=myquery, keys="\\\\demographics\\\\SEX\\\\")
## ...opps, added wrong term, lets delete it...
# hpds::query.crosscounts.delete(query=myquery, keys="\\\\demographics\\\\SEX\\\\")

}
