% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpds.R
\name{query.filter.add}
\alias{query.filter.add}
\title{Add a "Filter" restriction to a query instance.}
\usage{
query.filter.add(query, keys, verbose = FALSE)
}
\arguments{
\item{query}{A query instance object.}

\item{keys}{One or more keys to add to the given query object's filter list.}

\item{verbose}{Flag to display additional runtime information.}
}
\description{
Add a "Filter" restriction to a query instance.
}
