% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpds.R
\name{query.filter.delete}
\alias{query.filter.delete}
\title{Delete a "Filter" restriction from a query instance.}
\usage{
query.filter.delete(query, keys, verbose = FALSE)
}
\arguments{
\item{query}{A query instance object.}

\item{keys}{One or more keys to delete from the given query object's filter list.}

\item{verbose}{Flag to display additional runtime information.}
}
\description{
Delete a "Filter" restriction from a query instance.
}
\examples{

# myconn <- picsure::connect(url="http://your.server/PIC-SURE/", token="your-security-token")
# myres <- hpds::get.resource(connection=myconn, resourceUUID="YOUR-UUID-0000")
# myquery <- hpds::new.query(resource=myres)
# hpds::query.filter.add(query=myquery, keys="\\demographics\\SEX\\", values="Male")
## ...opps, added wrong term, lets delete it...
# hpds::query.filter.delete(query=myquery, keys="\\demographics\\SEX\\")

}
