% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpds.R
\name{query.from.uuid}
\alias{query.from.uuid}
\title{Retrieve an existing query by its unique identifier}
\usage{
query.from.uuid(resource, queryUUID, verbose = FALSE)
}
\arguments{
\item{resource}{A resource object that the returned query object will execute against.}

\item{uuid}{the valid UUID of a query to retrieve}
}
\description{
Retrieve an existing query by its unique identifier
}
\examples{

# myconn <- picsure::connect(url="http://your.server/PIC-SURE/", token="your-security-token")
# myres <- hpds::get.resource(connection=myconn, resourceUUID="YOUR-UUID-0000")
# myquery <- hpds::query.from.uuid(resource=myres, queryUUID="UUID" )

}
