% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpds.R
\name{query.load}
\alias{query.load}
\title{Load a query from a JSON-formated string.}
\usage{
query.load(query, query.def = "")
}
\arguments{
\item{query}{A JSON string that defines the query instance object.}

\item{verbose}{Flag to display additional runtime information.}
}
\description{
Load a query from a JSON-formated string.
}
\examples{

# myconn <- picsure::connect(url="http://your.server/PIC-SURE/", token="your-security-token")
# myres <- hpds::get.resource(connection=myconn, resourceUUID="YOUR-UUID-0000")
# myquery <- hpds::new.query(resource=myres)

## ...modify the query by adding search criteria... ##

#  myquery <- hpds::query.load(query=myquerydef)

}
