% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpds.R
\name{query.save}
\alias{query.save}
\title{Save a query in JSON-format with any restrictions that have been added to it.}
\usage{
query.save(query, result.type = "dataframe", verbose = FALSE)
}
\arguments{
\item{query}{A query instance object.}

\item{verbose}{Flag to display additional runtime information.}
}
\description{
Save a query in JSON-format with any restrictions that have been added to it.
}
\examples{

# myconn <- picsure::connect(url="http://your.server/PIC-SURE/", token="your-security-token")
# myres <- hpds::get.resource(connection=myconn, resourceUUID="YOUR-UUID-0000")
# myquery <- hpds::new.query(resource=myres)

## ...modify the query by adding search criteria... ##

# querydef <- hpds::query.save(query=myquery)

}
