% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpds.R
\name{query.select.add}
\alias{query.select.add}
\title{Add a "select" restriction to a query instance.}
\usage{
query.select.add(query, keys, verbose = FALSE)
}
\arguments{
\item{query}{A query instance object.}

\item{keys}{One or more keys to add to the given query object's select list.}

\item{verbose}{Flag to display additional runtime information.}
}
\description{
Add a "select" restriction to a query instance.
}
