% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{replaceClause}
\alias{replaceClause}
\title{Return a copy of a query with one sub-query swapped for another.}
\usage{
replaceClause(query, target, replacement)
}
\arguments{
\item{query}{A clause or clause-group handle (the query to edit).}

\item{target}{A clause or clause-group handle (the thing to replace).}

\item{replacement}{A clause or clause-group handle (the substitute).}
}
\value{
A new clause or clause-group handle.
}
\description{
Matching is structural (see [`removeSubQuery()`][picsure::removeSubQuery]).
}
\examples{
\dontrun{
adjusted <- picsure::replaceClause(full_query, old_age, new_age)
}
}
