% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpds.R
\name{resource.details}
\alias{resource.details}
\title{Get the details of a PIC-SURE resource.}
\usage{
resource.details(connection, resourceUUID, verbose = FALSE)
}
\arguments{
\item{connection}{A PIC-SURE connection object.}

\item{resourceUUID}{The UUID identity of a Resource hosted via the PIC-SURE connection.}

\item{verbose}{Flag to display additional runtime information.}
}
\value{
A string containing details of the requested Resource.
}
\description{
Get the details of a PIC-SURE resource.
}
