% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{showQuery}
\alias{showQuery}
\title{Prints the JSON representation of a query object}
\usage{
showQuery(query)
}
\arguments{
\item{query}{A query object}
}
\description{
Prints the JSON representation of a query object
}
\examples{

# query <- picsure::newQuery(session)
# query <- picsure::addClause(query, "\\demographics\\AGE\\", max = 10)
# picsure::showQuery(query)

}
