% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\docType{data}
\name{Zygosity}
\alias{Zygosity}
\title{Genotype codes for SNP / variant-spec genomic filters.}
\format{
A list of `picsure_enum_member` objects:
\describe{
  \item{`HETEROZYGOUS`}{Heterozygous (`0/1`).}
  \item{`HOMOZYGOUS`}{Homozygous (`1/1`).}
  \item{`HETEROZYGOUS_OR_HOMOZYGOUS`}{Either (`1/1,0/1`).}
}
}
\usage{
Zygosity
}
\description{
Pass a member to [`buildGenomicFilter()`][picsure::buildGenomicFilter]'s
`values` argument for a SNP variant-spec key. Mirrors Python's
`picsure.Zygosity`.
}
\examples{
\dontrun{
picsure::buildGenomicFilter(
  "chr5:148481541:T:A",
  values = picsure::Zygosity$HETEROZYGOUS
)
}
}
\keyword{datasets}
