% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc.R
\name{bdc.addClause}
\alias{bdc.addClause}
\title{Add a clause to a query}
\usage{
bdc.addClause(
  query,
  keys,
  type = "FILTER",
  min = NULL,
  max = NULL,
  categories = NULL
)
}
\arguments{
\item{query}{A query object}

\item{keys}{A key or list of keys to add a clause for}

\item{type}{One of: (“SELECT”, “REQUIRE”, “ANYOF”, “FILTER”)}

\item{min}{Optional parameter for continuous "FILTER" clauses. Specifies the minimum value}

\item{max}{Optional parameter for continuous "FILTER" clauses. Specifies the maximum value}

\item{categories}{Optional parameter for categorical "FILTER" clauses to match}
}
\value{
A query object
}
\description{
Add a clause to a query
}
\examples{

# query <- picsure::bdc.newQuery(session)
# query <- picsure::bdc.addClause(query, "\\phs000001\\pht000001\\phv00000001\\variable1\\", "FILTER", min = 1, max = 2)
# query <- picsure::bdc.addClause(query, "\\phs000001\\pht000001\\phv00000001\\variable2\\", "FILTER", categories = list("Yes", "Maybe"))
# count <- picsure::getCount(query)
# results <- picsure::getResults(query)

# query <- picsure::bdc.newQuery(session)
# query <- picsure::bdc.addClause(query, "\\phs000001\\pht000001\\phv00000001\\variable3\\", "SELECT")
# results <- picsure::getResults(query)

# query <- picsure::bdc.newQuery(session)
# query <- picsure::bdc.addClause(query, "\\phs000001\\pht000001\\phv00000001\\variable4\\", "REQUIRE")
# results <- picsure::getResults(query)

# query <- picsure::bdc.newQuery(session)
# query <- picsure::bdc.addClause(query, "\\phs000001\\pht000001\\phv00000001\\variable5\\", "ANYOF")
# results <- picsure::getResults(query)

}
