% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\docType{data}
\name{ClauseType}
\alias{ClauseType}
\title{Filter clause types.}
\format{
A list of `picsure_enum_member` objects:
\describe{
  \item{`FILTER`}{Filter by categorical values or numeric range.}
  \item{`ANYRECORD`}{Match records where the concept path *or any
    descendant* has a value (wire: `ANY_RECORD_OF`).}
  \item{`SELECT`}{Include the concept path(s) in query output.}
  \item{`REQUIRE`}{Require the concept path to have a non-null value
    (wire: `REQUIRED`).}
}
}
\usage{
ClauseType
}
\description{
Pass a member to [`createClause()`][picsure::createClause]'s `type`
argument. Mirrors Python's `picsure.ClauseType`.
}
\examples{
\dontrun{
picsure::createClause(
  "\\\\phs1\\\\pht1\\\\phv1\\\\sex\\\\",
  type = picsure::ClauseType$FILTER,
  categories = "male"
)
}
}
\keyword{datasets}
