% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{exportCSV}
\alias{exportCSV}
\title{Export query results to a CSV file.}
\usage{
exportCSV(session, query, path, ...)
}
\arguments{
\item{session}{A session object produced by [`connect()`][picsure::connect].}

\item{query}{A clause group from
[`buildClauseGroup()`][picsure::buildClauseGroup].}

\item{path}{Destination file path.}

\item{...}{Additional keyword arguments forwarded to the Python
`Session.exportPFB()` call.}
}
\value{
The path, invisibly.
}
\description{
Export query results to a CSV file.
}
\examples{
\dontrun{
picsure::exportCSV(bdc, full_query, "~/cohort.csv")
}
}
