% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpds.R
\name{extract.count}
\alias{extract.count}
\title{Extract the number of results in the given data dictionary lookup.}
\usage{
extract.count(dictionary.results, verbose = FALSE)
}
\arguments{
\item{dictionary.results}{A data dictionary search results object.}

\item{verbose}{Flag to display additional runtime information.}
}
\value{
An integer of how many data dictionary entries were found in
}
\description{
Extract the number of results in the given data dictionary lookup.
}
\examples{

# myconn <- picsure::connect(url="http://your.server/PIC-SURE/", token="your-security-token")
# myres <- hpds::get.resource(connection=myconn, resourceUUID="YOUR-UUID-0000")
# asthma.terms <- hpds::find.in.dictionary(resource=myres, term="asthma")
# extract.count(asthma.terms)

}
