% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpds.R
\name{extract.count}
\alias{extract.count}
\title{Extract the number of results in the given data dictionary lookup.}
\usage{
extract.count(dictionary.results, verbose = FALSE)
}
\arguments{
\item{dictionary.results}{A data dictionary search results object.}

\item{verbose}{Flag to display additional runtime information.}
}
\value{
An integer of how many data dictionary entries were found in
}
\description{
Extract the number of results in the given data dictionary lookup.
}
