% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpds.R
\name{extract.entries}
\alias{extract.entries}
\title{Extract the entries of all the results in a given data dictionary lookup.}
\usage{
extract.entries(dictionary.results, verbose = FALSE)
}
\arguments{
\item{dictionary.results}{A dictionary results object.}

\item{verbose}{Flag to display additional runtime information.}
}
\value{
A collection of result entries.
}
\description{
Extract the entries of all the results in a given data dictionary lookup.
}
\examples{

# myconn <- picsure::connect(url="http://your.pic-sure.server/PIC-SURE/", token="your-security-token")
# myres <- hpds::get.resource(connection=myconn, resourceUUID="612d00b7-YOUR-UUID-0000-1efe587f0bcd")
# asthma.terms <- hpds::find.in.dictionary(resource=myres, term="asthma")
# extract.entries(asthma.terms)

}
