% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facets.R
\name{facets}
\alias{facets}
\title{Build a FacetSet for narrowing search results.}
\usage{
facets(session)
}
\arguments{
\item{session}{A session object produced by [`connect()`][picsure::connect].}
}
\value{
An opaque FacetSet handle.
}
\description{
Returns an opaque `FacetSet` handle tied to this session. Add entries with
[`addFacet()`][picsure::addFacet] and remove them with
[`removeFacet()`][picsure::removeFacet]. Pass the final FacetSet to
[`searchDictionary()`][picsure::searchDictionary] via the `facets` argument.
}
\examples{
\dontrun{
fs <- picsure::facets(bdc)
fs <- picsure::addFacet(fs, "study_ids", "phs000007")
picsure::searchDictionary(bdc, "sex", facets = fs)
}
}
