% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpds.R
\name{find.in.dictionary}
\alias{find.in.dictionary}
\title{Search the data dictionary of a PIC-SURE resource.}
\usage{
find.in.dictionary(resource, term, verbose = FALSE)
}
\arguments{
\item{resource}{A PIC-SURE resource object.}

\item{term}{A string to search for in the resource's data dictionary.}

\item{verbose}{Flag to display additional runtime information.}
}
\value{
An object representing the search results.
}
\description{
Search the data dictionary of a PIC-SURE resource.
}
