% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{getResultByQueryUUID}
\alias{getResultByQueryUUID}
\title{Executes a query with a given queryUUID}
\usage{
getResultByQueryUUID(session, queryUUID)
}
\arguments{
\item{query}{A query object}

\item{resultType}{The result type to query for:
"COUNT" queries only for the number of results
"DATA_FRAME" queries for the entire result set}
}
\value{
The result of the query
}
\description{
Executes a query with a given queryUUID
}
\examples{

# count <- picsure::runQuery(query, "COUNT")
# results <- picsure::runQuery(query, "DATA_FRAME")

}
