% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\docType{data}
\name{GroupOperator}
\alias{GroupOperator}
\title{Logical operators for combining clauses in a group.}
\format{
A list of `picsure_enum_member` objects:
\describe{
  \item{`AND`}{All clauses must match.}
  \item{`OR`}{At least one clause must match.}
}
}
\usage{
GroupOperator
}
\description{
Pass a member to [`buildQuery()`][picsure::buildQuery]'s
`operator` argument. Mirrors Python's `picsure.GroupOperator`.
}
\examples{
\dontrun{
c1 <- picsure::createSubQuery("\\\\phs1\\\\sex\\\\",
                             type = picsure::ClauseType$FILTER,
                             categories = "male")
c2 <- picsure::createSubQuery("\\\\phs1\\\\copd\\\\",
                             type = picsure::ClauseType$FILTER,
                             categories = "Yes")
picsure::buildQuery(
  list(c1, c2),
  operator = picsure::GroupOperator$AND
)
}
}
\keyword{datasets}
