% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6-classes.R
\docType{class}
\name{HpdsAttribListKeyValues}
\alias{HpdsAttribListKeyValues}
\title{R6 class used to store/manipulate Key+Value query parameter lists - DO NOT CREATE THIS OBJECT DIRECTLY!}
\format{\code{\link{HpdsAttribListKeyValues}} object.}
\usage{
HpdsAttribListKeyValues
}
\value{
Object of \code{\link{R6Class}} used to access a HPDS-hosted resource's data dictionary.
}
\description{
R6 class used to store/manipulate Key+Value query parameter lists - DO NOT CREATE THIS OBJECT DIRECTLY!
}
\section{Methods}{

\describe{
  \item{Documentation}{For full documentation of each method go to https://github.com/hms-dbmi/pic-sure-r-adapter-hpds}
  \item{\code{new(inst_list, help_text)}}{This method is used to create new object of this class. DO NOT CREATE THIS OBJECT DIRECTLY!}

  \item{\code{add()}}{This method adds one or more entries to the query parameter list.}
  \item{\code{delete(keys)}}{This method deletes one or more entries from the query parameter list.}
  \item{\code{clear()}}{This method clears all entries from the query parameter list.}
  \item{\code{show()}}{This method displays the entries of the query parameter list.}
  \item{\code{getQueryValues()}}{This is an internally used method that returns the entries for use by the parent query object.}}
}

\keyword{data}
