% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpds.R
\name{new.query}
\alias{new.query}
\title{Create a new query instance with no restrictions.}
\usage{
new.query(resource, verbose = FALSE)
}
\arguments{
\item{resource}{A resource object that the returned query object will execute against.}

\item{verbose}{Flag to display additional runtime information.}
}
\description{
Create a new query instance with no restrictions.
}
