% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{picsure-package}
\alias{picsure}
\alias{picsure-package}
\title{picsure: R Client for the PIC-SURE API}
\description{
Thin R wrapper over the PIC-SURE Python adapter. Provides functions to connect to PIC-SURE networks, search the data dictionary, build nested query clauses, run queries, and export results. Python interop is handled transparently via reticulate; the required Python environment is provisioned automatically on first use.
}
\author{
\strong{Maintainer}: PIC-SURE Team \email{pic-sure-team@hms.harvard.edu}

}
\keyword{internal}
