% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6-classes.R
\docType{class}
\name{PicSureHpdsDictionaryResult}
\alias{PicSureHpdsDictionaryResult}
\title{R6 class contain the results of a search against a HPDS resource's data dictionary - DO NOT CREATE THIS OBJECT DIRECTLY!}
\format{\code{\link{PicSureHpdsDictionaryResult}} object.}
\usage{
PicSureHpdsDictionaryResult
}
\value{
Object of \code{\link{R6Class}} used to access a HPDS-hosted resource's data dictionary.
}
\description{
R6 class contain the results of a search against a HPDS resource's data dictionary - DO NOT CREATE THIS OBJECT DIRECTLY!
}
\keyword{data}
