% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6-classes.R
\docType{class}
\name{PicSureHpdsQuery}
\alias{PicSureHpdsQuery}
\title{R6 class used to build a multi-use query to search against a HPDS resource's data - DO NOT CREATE THIS OBJECT DIRECTLY!}
\format{
\code{\link{PicSureHpdsQuery}} object.
}
\value{
Object of \code{\link{R6Class}} used to access a HPDS-hosted resource's data dictionary.
}
\description{
R6 class used to build a multi-use query to search against a HPDS resource's data - DO NOT CREATE THIS OBJECT DIRECTLY!

R6 class used to build a multi-use query to search against a HPDS resource's data - DO NOT CREATE THIS OBJECT DIRECTLY!
}
\section{Methods}{

\describe{
  \item{Documentation}{For full documentation of each method go to https://github.com/hms-dbmi/pic-sure-r-adapter-hpds}
  \item{\code{new(connection)}}{This method is used to create new object of this class. DO NOT CREATE THIS OBJECT DIRECTLY!}

  \item{\code{show()}}{This method displays a list of all settings specified for the query.}
  \item{\code{select()}}{This method accesses a reference to a query parameter list.}
  \item{\code{crosscounts()}}{This method accesses a reference to a query parameter list.}
  \item{\code{require()}}{This method accesses a reference to a query parameter list.}
  \item{\code{anyof()}}{This method accesses a reference to a query parameter list.}
  \item{\code{filter()}}{This method accesses a reference to a query parameter list.}
  \item{\code{getCount()}}{This method returns a count of how many records are discovered by the query.}
  \item{\code{getResults()}}{This method returns the records discovered by the query.}
  \item{\code{getResultsDataFrame()}}{This method returns the discovered records in a dataframe format.}
  \item{\code{getVariantsApproximateCount()}}{This method returns the a count of how many variant records are discovered by the query.}
  \item{\code{getVariantsDataFrame()}}{This method returns the discovered variant records in a dataframe format.}
  \item{\code{getRunDetails()}}{This method returns information the performance of the query.}}
}

\keyword{data}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PicSureHpdsQuery$new()}}
\item \href{#method-show}{\code{PicSureHpdsQuery$show()}}
\item \href{#method-select}{\code{PicSureHpdsQuery$select()}}
\item \href{#method-crosscounts}{\code{PicSureHpdsQuery$crosscounts()}}
\item \href{#method-require}{\code{PicSureHpdsQuery$require()}}
\item \href{#method-anyof}{\code{PicSureHpdsQuery$anyof()}}
\item \href{#method-filter}{\code{PicSureHpdsQuery$filter()}}
\item \href{#method-getCount}{\code{PicSureHpdsQuery$getCount()}}
\item \href{#method-getResults}{\code{PicSureHpdsQuery$getResults()}}
\item \href{#method-getResultsDataFrame}{\code{PicSureHpdsQuery$getResultsDataFrame()}}
\item \href{#method-getVariantsApproximateCount}{\code{PicSureHpdsQuery$getVariantsApproximateCount()}}
\item \href{#method-getVariantsDataFrame}{\code{PicSureHpdsQuery$getVariantsDataFrame()}}
\item \href{#method-getRunDetails}{\code{PicSureHpdsQuery$getRunDetails()}}
\item \href{#method-save}{\code{PicSureHpdsQuery$save()}}
\item \href{#method-load}{\code{PicSureHpdsQuery$load()}}
\item \href{#method-loadInternal}{\code{PicSureHpdsQuery$loadInternal()}}
\item \href{#method-buildQuery}{\code{PicSureHpdsQuery$buildQuery()}}
\item \href{#method-clone}{\code{PicSureHpdsQuery$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PicSureHpdsQuery$new(connection)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-show"></a>}}
\if{latex}{\out{\hypertarget{method-show}{}}}
\subsection{Method \code{show()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PicSureHpdsQuery$show()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-select"></a>}}
\if{latex}{\out{\hypertarget{method-select}{}}}
\subsection{Method \code{select()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PicSureHpdsQuery$select()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crosscounts"></a>}}
\if{latex}{\out{\hypertarget{method-crosscounts}{}}}
\subsection{Method \code{crosscounts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PicSureHpdsQuery$crosscounts()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-require"></a>}}
\if{latex}{\out{\hypertarget{method-require}{}}}
\subsection{Method \code{require()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PicSureHpdsQuery$require()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-anyof"></a>}}
\if{latex}{\out{\hypertarget{method-anyof}{}}}
\subsection{Method \code{anyof()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PicSureHpdsQuery$anyof()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-filter"></a>}}
\if{latex}{\out{\hypertarget{method-filter}{}}}
\subsection{Method \code{filter()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PicSureHpdsQuery$filter()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getCount"></a>}}
\if{latex}{\out{\hypertarget{method-getCount}{}}}
\subsection{Method \code{getCount()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PicSureHpdsQuery$getCount(asAsync = FALSE, timeout = 30)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getResults"></a>}}
\if{latex}{\out{\hypertarget{method-getResults}{}}}
\subsection{Method \code{getResults()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PicSureHpdsQuery$getResults(asAsync = FALSE, timeout = 30)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getResultsDataFrame"></a>}}
\if{latex}{\out{\hypertarget{method-getResultsDataFrame}{}}}
\subsection{Method \code{getResultsDataFrame()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PicSureHpdsQuery$getResultsDataFrame(asAsync = FALSE, timeout = 30)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getVariantsApproximateCount"></a>}}
\if{latex}{\out{\hypertarget{method-getVariantsApproximateCount}{}}}
\subsection{Method \code{getVariantsApproximateCount()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PicSureHpdsQuery$getVariantsApproximateCount(asAsync = FALSE, timeout = 30)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getVariantsDataFrame"></a>}}
\if{latex}{\out{\hypertarget{method-getVariantsDataFrame}{}}}
\subsection{Method \code{getVariantsDataFrame()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PicSureHpdsQuery$getVariantsDataFrame(asAsync = FALSE, timeout = 30)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getRunDetails"></a>}}
\if{latex}{\out{\hypertarget{method-getRunDetails}{}}}
\subsection{Method \code{getRunDetails()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PicSureHpdsQuery$getRunDetails()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-save"></a>}}
\if{latex}{\out{\hypertarget{method-save}{}}}
\subsection{Method \code{save()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PicSureHpdsQuery$save(resultType = "COUNT")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-load"></a>}}
\if{latex}{\out{\hypertarget{method-load}{}}}
\subsection{Method \code{load()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PicSureHpdsQuery$load(queryStr = "{}", merge = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-loadInternal"></a>}}
\if{latex}{\out{\hypertarget{method-loadInternal}{}}}
\subsection{Method \code{loadInternal()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PicSureHpdsQuery$loadInternal(queryObj, merge = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-buildQuery"></a>}}
\if{latex}{\out{\hypertarget{method-buildQuery}{}}}
\subsection{Method \code{buildQuery()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PicSureHpdsQuery$buildQuery(resultType = "COUNT")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PicSureHpdsQuery$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
