% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6-classes.R
\docType{class}
\name{PicSureHpdsResourceConnection}
\alias{PicSureHpdsResourceConnection}
\title{R6 class that allows access to the data dictionary and query services of a selected HPDS-hosted resources on a PIC-SURE network.}
\format{\code{\link{PicSureHpdsResourceConnection}} object.}
\usage{
PicSureHpdsResourceConnection
}
\value{
Object of \code{\link{R6Class}} used to access the dictionary and query services via the objects it returns.
}
\description{
R6 class that allows access to the data dictionary and query services of a selected HPDS-hosted resources on a PIC-SURE network.
}
\section{Methods}{

\describe{
  \item{Documentation}{For full documentation of each method go to https://github.com/hms-dbmi/pic-sure-r-adapter-hpds}
  \item{\code{new(connection, resource_uuid)}}{This method is used to create new object of this class which uses the passed PicSureConnection object for communication with the PIC-SURE network along with a UUID to identify a HPDS-hosted resource.}

  \item{\code{dictionary()}}{This method returns a \code{PicSureHpdsDictionary} object which is used to run lookups against a resources data dictionary.}
  \item{\code{query()}}{This method returns a new \code{PicSureHpdsQuery} object configured to run all commands against the previously specified HPDS-hosted resource.}}
}

\keyword{data}
