% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpds.R
\name{query.crosscounts.delete}
\alias{query.crosscounts.delete}
\title{Delete a "cross-counts" restriction from a query instance.}
\usage{
query.crosscounts.delete(query, keys, verbose = FALSE)
}
\arguments{
\item{query}{A query instance object.}

\item{keys}{One or more keys to delete from the given query object's cross-count list.}

\item{verbose}{Flag to display additional runtime information.}
}
\description{
Delete a "cross-counts" restriction from a query instance.
}
