% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpds.R
\name{query.filter.delete}
\alias{query.filter.delete}
\title{Delete a "Filter" restriction from a query instance.}
\usage{
query.filter.delete(query, keys, verbose = FALSE)
}
\arguments{
\item{query}{A query instance object.}

\item{keys}{One or more keys to delete from the given query object's filter list.}

\item{verbose}{Flag to display additional runtime information.}
}
\description{
Delete a "Filter" restriction from a query instance.
}
