% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpds.R
\name{query.run}
\alias{query.run}
\title{Run a query instance using any restrictions that have been added to it.}
\usage{
query.run(query, result.type = "dataframe", verbose = FALSE)
}
\arguments{
\item{query}{A query instance object.}

\item{result.type}{A string specifying what type of results to return. Possible values: "count", "results", "dataframe" and "crosscount".}

\item{verbose}{Flag to display additional runtime information.}
}
\description{
Run a query instance using any restrictions that have been added to it.
}
