% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpds.R
\name{query.show}
\alias{query.show}
\title{Prints to screen the contents of a query instance's restriction parameters.}
\usage{
query.show(query, verbose = FALSE)
}
\arguments{
\item{query}{A query instance object.}

\item{verbose}{Flag to display additional runtime information.}
}
\description{
Prints to screen the contents of a query instance's restriction parameters.
}
