% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{runQuery}
\alias{runQuery}
\title{Executes a query}
\usage{
runQuery(query, resultType = NULL)
}
\arguments{
\item{query}{A query object}

\item{resultType}{The result type to query for:
"COUNT" queries only for the number of results
"DATA_FRAME" queries for the entire result set}
}
\value{
The result of the query
}
\description{
Executes a query
}
\examples{

# count <- picsure::runQuery(query, "COUNT")
# results <- picsure::runQuery(query, "DATA_FRAME")

}
