% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\docType{data}
\name{VariantFrequency}
\alias{VariantFrequency}
\title{Variant population-frequency buckets.}
\format{
A list of `picsure_enum_member` objects:
\describe{
  \item{`RARE`}{Rare variants.}
  \item{`COMMON`}{Common variants.}
  \item{`NOVEL`}{Novel variants.}
}
}
\usage{
VariantFrequency
}
\description{
Pass a member to [`buildGenomicFilter()`][picsure::buildGenomicFilter]'s
`values` argument for the `"Variant_frequency_as_text"` key. Mirrors
Python's `picsure.VariantFrequency`.
}
\examples{
\dontrun{
picsure::buildGenomicFilter(
  "Variant_frequency_as_text",
  values = picsure::VariantFrequency$RARE
)
}
}
\keyword{datasets}
